test: move sklearn-dependent router test to integration (fix CI test-unit)
CI `test-unit` runs [dev]-only with --disable-socket; scikit-learn comes via [ml],
so the joblib+DummyClassifier test failed to import sklearn there. Moved it to
tests/integration/search/ (importorskip sklearn/joblib). The env-safe router
fallbacks (corrupt-model, predict-raises) stay as unit tests.
